17 chat and user db

+++ src/nestjs/src/modules/database/user/controller.ts
@@ -11,14 +11,14 @@ import {
 import { Response } from "express";
 
 import { UserService } from "./service";
-import { UserPost } from "./dto";
+import { UserPost, UserInfoPost } from "./dto";
 
 @Controller("user")
 export class UserController {
 	constructor(private readonly userService: UserService) {}
 
 	@Post()
-	create(@Body() post: UserPost) {
+	async create(@Body() post: UserPost) {
 		return this.userService.create(post);
 	}
 
@@ -35,14 +35,18 @@ export class UserController {
 	@Put(":id")
 	async update(
 		@Param("id") userId: number,
-		@Body() userPost: UserPost,
+		@Body() userPost: UserInfoPost,
 		@Res() res: Response,
 	) {
 		res.send(await this.userService.update(userId, userPost));
 	}
-
+	
 	@Delete(":id")
 	async delete(@Param("id") userId: number, @Res() res: Response) {
-		res.send(await this.userService.delete(userId));
+		const user = await this.userService.delete(userId);
+		if (user)
+			res.send(await this.userService.delete(userId));
+		else
+			console.log('in delete user nb : ', userId, ' not found')
 	}
 }
